% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_tree.R
\name{classification_tree}
\alias{classification_tree}
\title{Classification Tree}
\usage{
classification_tree(x, y, control = rpart::rpart.control(), ...)
}
\arguments{
\item{x}{A data frame containing the predictor variables.}

\item{y}{A factor vector containing the categorical response variable.}

\item{control}{A list of control parameters for the rpart function (optional).}

\item{...}{Additional arguments to be passed to the rpart function.}
}
\value{
A fitted rpart object.
}
\description{
Fit a classification tree model using the rpart package.
}
\examples{
data(iris)
x <- iris[, -5]
y <- iris$Species
class_tree <- classification_tree(x, y)
summary(class_tree)
}
