% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_tree.R
\name{performance_tree}
\alias{performance_tree}
\title{Evaluate Tree Performance}
\usage{
performance_tree(tree, x, y, plot = TRUE)
}
\arguments{
\item{tree}{A fitted rpart object.}

\item{x}{A data frame containing the predictor variables.}

\item{y}{A vector containing the response variable (either continuous or categorical).}

\item{plot}{A logical value indicating whether to plot the tree (TRUE) or not (FALSE). Default is TRUE.}
}
\value{
A list containing the performance metrics.
}
\description{
Create performance evaluation tables and plots for a classification or regression tree.
}
\examples{
data(mtcars)
x <- mtcars[, -1]
y <- mtcars$mpg
reg_tree <- regression_tree(x, y)
performance <- performance_tree(reg_tree, x, y)
}
