% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tree.R
\name{predict_tree}
\alias{predict_tree}
\title{Predict with Tree}
\usage{
predict_tree(tree, newdata, ...)
}
\arguments{
\item{tree}{A fitted rpart object.}

\item{newdata}{A data frame containing the predictor variables for new observations.}

\item{...}{Additional arguments to be passed to the predict.rpart function.}
}
\value{
A vector of predicted values.
}
\description{
Make predictions using a fitted classification or regression tree.
}
\examples{
data(mtcars)
x <- mtcars[, -1]
y <- mtcars$mpg
reg_tree <- regression_tree(x, y)
predictions <- predict_tree(reg_tree, x)
print(predictions)
}
