% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_tree.R
\name{regression_tree}
\alias{regression_tree}
\title{Regression Tree}
\usage{
regression_tree(x, y, control = rpart::rpart.control(), ...)
}
\arguments{
\item{x}{A data frame containing the predictor variables.}

\item{y}{A vector containing the continuous response variable.}

\item{control}{A list of control parameters for the rpart function (optional).}

\item{...}{Additional arguments to be passed to the rpart function.}
}
\value{
A fitted rpart object.
}
\description{
Fit a regression tree model using the rpart package.
}
\examples{
data(mtcars)
x <- mtcars[, -1]
y <- mtcars$mpg
reg_tree <- regression_tree(x, y)
}
